moved get place detail into landmark
