+++ app/src/AREALITY/app/src/main/AndroidManifest.xml
@@ -21,15 +21,16 @@
         <activity android:name=".MainActivity"></activity>
 
         <activity android:name=".LandmarkPage">
-
-        </activity>
-        <activity android:name=".MyGLActivity"></activity>
-
-        <activity android:name=".MapsActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity android:name=".MyGLActivity"></activity>
+
+        <activity android:name=".MapsActivity">
+
+
         </activity>
       
         <activity android:name=".LoginActivity" />
