+++ app/src/AREALITY/app/src/main/java/com/example/areality/LandmarkPage.java
@@ -2,7 +2,16 @@ package com.example.areality;
 
 import android.app.Activity;
 import android.os.Bundle;
+import android.os.StrictMode;
+import android.util.Log;
 import android.widget.TextView;
+import android.widget.Toast;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
 
 public class LandmarkPage extends Activity {
 
@@ -12,8 +21,16 @@ public class LandmarkPage extends Activity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_my_gl);
 
-//        glView = new MyGLSurfaceView(this);           // Allocate a GLSurfaceView
-//        setContentView(glView);
+        String testPlaceId = "ChIJIQBpAG2ahYAR_6128GcTUEo";
+        String urlString = getDetailUrl(testPlaceId);
+        String result = "didn't work";
+
+        try {
+            result = makeHTTPRequest(urlString);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
 
         setContentView(R.layout.activity_landmark_page);
         glView = (MyGLSurfaceView) findViewById(R.id.glsurfaceview);
@@ -22,7 +39,9 @@ public class LandmarkPage extends Activity {
                 "2. Does line breaks work" +
                 "3. ");
 
-//        glView.setMyGLRenderer(this);
+
+//        Toast.makeText(LandmarkPage.this,result.toString(), Toast.LENGTH_LONG).show();
+
     }
     private void setInfo(String message){
         TextView textView = (TextView) findViewById(R.id.textView);
@@ -40,5 +59,36 @@ public class LandmarkPage extends Activity {
     protected void onResume() {
         super.onResume();
         glView.onResume();
+    }
+
+    private String getDetailUrl(String placeId) {
+
+        StringBuilder googlePlacesDetailUrl = new StringBuilder("https://maps.googleapis.com/maps/api/place/details/json?");
+        googlePlacesDetailUrl.append("placeid=" + placeId);
+        googlePlacesDetailUrl.append("&key=" + "AIzaSyD3FM6gEwhGLsi8ig7ebIZr4g46RgkrnQQ");
+        Log.d("getUrl", googlePlacesDetailUrl.toString());
+        return (googlePlacesDetailUrl.toString());
+    }
+
+    protected String makeHTTPRequest(String urlString) throws IOException {
+        StrictMode.ThreadPolicy policy = new StrictMode.ThreadPolicy.Builder().permitAll().build();
+        StrictMode.setThreadPolicy(policy);
+        HttpURLConnection connection;
+        try {
+            URL url = new URL(urlString);
+            connection = (HttpURLConnection) url.openConnection();
+            connection.setRequestMethod("GET");
+            connection.setDoOutput(true);
+            connection.connect();
+            BufferedReader rd = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+            String content = "", line;
+            while ((line = rd.readLine()) != null) {
+                content += line + "\n";
+            }
+            return content;
+        } catch (IOException e) {
+            Log.d("error", e.toString());
+            return e.toString();
+        }
     }
 }
